changed nip59 custom code with nostr-sdk - To be tested
